days commit

+++ bookmanagement/src/main/java/com/zerp/bookmanagement/Repository/BookRepository.java
@@ -4,11 +4,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.zerp.bookmanagement.Model.Book;
+
+import java.math.BigInteger;
 import java.util.List;
 
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, BigInteger> {
 
     List<Book> findByAuthor(String author);
